se acomodo el menu temporal proteger rutas laravel y algunos cambios de titulos

+++ routes/web.php
@@ -17,10 +17,10 @@ Route::post('/login', 'App\Http\Controllers\Auth\LoginController@login');
 Route::post('/logout', 'App\Http\Controllers\Auth\LoginController@logout');
 Route::post('/registrar', 'App\Http\Controllers\Auth\RegisterController@register');
 //para admin de usuarios
-Route::get('/users', 'App\Http\Controllers\Admin\UserController@index');
-Route::get('/users/{user}/edit', 'App\Http\Controllers\Admin\UserController@edit');
-Route::put('/users/{user}', 'App\Http\Controllers\Admin\UserController@update');
-Route::delete('/users/{user}', 'App\Http\Controllers\Admin\UserController@destroy');
+Route::get('/users', 'App\Http\Controllers\Admin\UserController@index')->middleware('can:admin.user.index');
+Route::get('/users/{user}/edit', 'App\Http\Controllers\Admin\UserController@edit')->middleware('can:admin.user.edit');
+Route::put('/users/{user}', 'App\Http\Controllers\Admin\UserController@update')->middleware('can:admin.user.edit');;
+Route::delete('/users/{user}', 'App\Http\Controllers\Admin\UserController@destroy')->middleware('can:admin.user.destroy');;
 
 //Route::resource('users', UserController::class);
